refactor directory and moving .mlmodel files
