--- tools/core_ml/__init__.py
@@ -1,1 +1,0 @@
-from _mxnet_converter import *
